Expanded on the core behaviours.
The Trader now starts with a required amount of units and a maximum buy price. It will request an offer from each other agent and accept the offer with the lowest price per unit if the price is lowest than the maximum buy price. Each time an offer is accepted, the required units will be updated. When the required units reaches zero the agent stops requesting offers.

Each Retailer starts with a number of units held and a price range. It will send offers based on these values. The units held is updated when an offer is accepted. If the Retailed no longer holds any units, it will send a failure message when queried.